// my-turborepo/packages/yc-models/src/core/context/types.ts
/**
 * All of the possible endpoints
 */
export enum Endpoints {
  ALL = "all", // All endpoints
  NETWORKS = "networks",
  TOKENS = "tokens",
  USERS = "users",
  STRATEGIES = "strateges",
  PROTOCOLS = "protocols",
  FLOWS = "flows",
  FUNCTIONS = "functions",
  ARGUMENTS = "arguments",
  ADDRESSES = "addresses",
  ACTIONS = "actions",
}